+++ src/build_number.h
@@ -1,1 +1,1 @@
-#define BUILD 495
+#define BUILD 496